KeySelection: move text selection from readerhighlight to readerkeyselection
